Script to calculate the spreads for USDT trading pairs
