--- gateway-files/logs/logs_gateway_app.log.2025-11-28
@@ -1,45 +1,0 @@
-2025-11-27 18:39:38 | info | 	⚡️ Gateway version 2.10.0 starting at http://localhost:15888
-2025-11-27 18:39:38 | info | 	🔧 Log level configured as: debug
-2025-11-27 18:39:38 | info | 	Checking for processes using port 15888...
-2025-11-27 18:39:38 | info | 	No process found using port 15888
-2025-11-27 18:39:38 | info | 	🔴 Running in development mode with (unsafe!) HTTP endpoints
-2025-11-27 18:39:38 | info | 	📓 Documentation available at http://localhost:15888/docs
-2025-11-27 18:40:06 | info | 	⚡️ Gateway version 2.10.0 starting at http://localhost:15888
-2025-11-27 18:40:06 | info | 	🔧 Log level configured as: debug
-2025-11-27 18:40:06 | info | 	Checking for processes using port 15888...
-2025-11-27 18:40:06 | info | 	No process found using port 15888
-2025-11-27 18:40:06 | info | 	🔴 Running in development mode with (unsafe!) HTTP endpoints
-2025-11-27 18:40:07 | info | 	📓 Documentation available at http://localhost:15888/docs
-2025-11-27 18:40:21 | info | 	Getting available DEX connectors and networks
-2025-11-27 18:40:21 | info | 	Available connectors: jupiter, meteora, raydium, uniswap, 0x, pancakeswap, pancakeswap-sol
-2025-11-27 18:40:21 | info | 	Getting available blockchain networks
-2025-11-27 18:40:21 | info | 	Available chains: ethereum (9 networks), solana (2 networks)
-2025-11-27 18:40:21 | info | 	Getting all configuration namespaces
-2025-11-27 18:40:21 | info | 	Found 23 namespaces: 0x, ethereum, ethereum-arbitrum, ethereum-avalanche, ethereum-base, ethereum-bsc, ethereum-celo, ethereum-mainnet, ethereum-optimism, ethereum-polygon, ethereum-sepolia, helius, infura, jupiter, meteora, pancakeswap, pancakeswap-sol, raydium, server, solana, solana-devnet, solana-mainnet-beta, uniswap
-2025-11-27 18:40:21 | info | 	Getting all configurations
-2025-11-27 19:05:32 | info | 	Getting available DEX connectors and networks
-2025-11-27 19:05:32 | info | 	Available connectors: jupiter, meteora, raydium, uniswap, 0x, pancakeswap, pancakeswap-sol
-2025-11-27 19:05:32 | info | 	Getting available blockchain networks
-2025-11-27 19:05:32 | info | 	Available chains: ethereum (9 networks), solana (2 networks)
-2025-11-27 19:05:32 | info | 	Getting all configuration namespaces
-2025-11-27 19:05:32 | info | 	Found 23 namespaces: 0x, ethereum, ethereum-arbitrum, ethereum-avalanche, ethereum-base, ethereum-bsc, ethereum-celo, ethereum-mainnet, ethereum-optimism, ethereum-polygon, ethereum-sepolia, helius, infura, jupiter, meteora, pancakeswap, pancakeswap-sol, raydium, server, solana, solana-devnet, solana-mainnet-beta, uniswap
-2025-11-27 19:05:32 | info | 	Getting all configurations
-2025-11-27 20:31:11 | info | 	⚡️ Gateway version 2.10.0 starting at http://localhost:15888
-2025-11-27 20:31:11 | info | 	🔧 Log level configured as: debug
-2025-11-27 20:31:11 | info | 	Checking for processes using port 15888...
-2025-11-27 20:31:11 | info | 	No process found using port 15888
-2025-11-27 20:31:11 | info | 	🔴 Running in development mode with (unsafe!) HTTP endpoints
-2025-11-27 20:31:12 | info | 	📓 Documentation available at http://localhost:15888/docs
-2025-11-27 20:31:24 | info | 	⚡️ Gateway version 2.10.0 starting at http://localhost:15888
-2025-11-27 20:31:24 | info | 	🔧 Log level configured as: debug
-2025-11-27 20:31:24 | info | 	Checking for processes using port 15888...
-2025-11-27 20:31:24 | info | 	No process found using port 15888
-2025-11-27 20:31:24 | info | 	🔴 Running in development mode with (unsafe!) HTTP endpoints
-2025-11-27 20:31:24 | info | 	📓 Documentation available at http://localhost:15888/docs
-2025-11-27 20:31:35 | info | 	Getting available DEX connectors and networks
-2025-11-27 20:31:35 | info | 	Available connectors: jupiter, meteora, raydium, uniswap, 0x, pancakeswap, pancakeswap-sol
-2025-11-27 20:31:35 | info | 	Getting available blockchain networks
-2025-11-27 20:31:35 | info | 	Available chains: ethereum (9 networks), solana (2 networks)
-2025-11-27 20:31:35 | info | 	Getting all configuration namespaces
-2025-11-27 20:31:35 | info | 	Found 23 namespaces: 0x, ethereum, ethereum-arbitrum, ethereum-avalanche, ethereum-base, ethereum-bsc, ethereum-celo, ethereum-mainnet, ethereum-optimism, ethereum-polygon, ethereum-sepolia, helius, infura, jupiter, meteora, pancakeswap, pancakeswap-sol, raydium, server, solana, solana-devnet, solana-mainnet-beta, uniswap
-2025-11-27 20:31:35 | info | 	Getting all configurations
